Moved reusable components to common folder
