+++ app/src/pages/inside/common/attachmentList/attachmentsList.scss
