+++ app/src/pages/inside/common/attachmentList/types.ts
@@ -1,0 +1,8 @@
+export interface Attachment {
+  fileName: string;
+  fileSize: number;
+  id: number;
+  fileType: string;
+  src?: string;
+  hasThumbnail?: boolean;
+}
