+++ app/src/pages/inside/testCaseLibraryPage/createTestCaseModal/testCaseDetails/steps/step/step.tsx
@@ -20,8 +20,7 @@ import { isEmpty } from 'es-toolkit/compat';
 
 import { createClassnames } from 'common/utils';
 import { FieldErrorHint, FieldProvider } from 'components/fields';
-import { Attachment } from 'pages/inside/testCaseLibraryPage/types';
-import { AttachmentList } from 'pages/inside/testCaseLibraryPage/attachmentList';
+import { AttachmentList, type Attachment } from 'pages/inside/common/attachmentList';
 import { FieldSection } from 'pages/inside/common/fieldSection';
 
 import styles from './step.scss';
