+++ app/src/pages/inside/testCaseLibraryPage/testCaseDetailsPage/precondition/precondition.tsx
@@ -20,8 +20,7 @@ import { isEmpty } from 'es-toolkit/compat';
 import { createClassnames } from 'common/utils';
 import { commonMessages } from 'pages/inside/common/common-messages';
 
-import { AttachmentList } from '../../attachmentList';
-import { Attachment } from '../../types';
+import { AttachmentList, type Attachment } from '../../../common/attachmentList';
 
 import styles from './preconditions.scss';
 
@@ -41,23 +40,21 @@ export const Precondition = ({ preconditions }: PreconditionProps) => {
     <section className={cx('precondition')}>
       <div className={cx('precondition__type')}>P</div>
       <div className={cx('precondition__content')}>
-        {
-          preconditions.value &&
-            <div className={cx('precondition__value')}>
-              {preconditions.value}
-            </div>
-        }
-        {
-          !isEmpty(preconditions.attachments) &&
-            <div className={cx('precondition__attachments-wrapper')}>
-              <h4>{formatMessage(commonMessages.attachments)} {preconditions.attachments.length}</h4>
-              <AttachmentList
-                attachments={preconditions.attachments}
-                className={cx('precondition__attachments-list')}
-                withPreview
-              />
-            </div>
-        }
+        {preconditions.value && (
+          <div className={cx('precondition__value')}>{preconditions.value}</div>
+        )}
+        {!isEmpty(preconditions.attachments) && (
+          <div className={cx('precondition__attachments-wrapper')}>
+            <h4>
+              {formatMessage(commonMessages.attachments)} {preconditions.attachments.length}
+            </h4>
+            <AttachmentList
+              attachments={preconditions.attachments}
+              className={cx('precondition__attachments-list')}
+              withPreview
+            />
+          </div>
+        )}
       </div>
     </section>
   );
